requirements/base.txt
tornado==4.1
futures==3.0.2
redis==2.10.3
pymongo==2.8
statsd==3.1
pytz==2013.7
requests==2.7.0
lxml==3.4.4
jsonpath-rw==1.4.0
python-dateutil==2.4.2
Pygments==2.0.2
dogpile.cache==0.5.6
matchers==0.22
PyYAML==3.11
# performance
plop==0.1.1
yappi==0.94
Sphinx==1.3.1
